Remove AbstractStoreFileManager and make minor changes in PersistedStoreFileManager and DefaultStoreFileManager
